% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_clean}
\alias{data_clean}
\title{FWS data on 7yr ESA implementation plan}
\format{A data frame with 363 rows and 10 variables
\describe{
\item{\code{Package.Name}}{Species grouping}
\item{\code{Species.name}}{Name of species; common or scientific}
\item{\code{Action.Type}}{Type of action}
\item{\code{Lead.RO}}{Primary FWS regional office responsible for species assessment}
\item{\code{Priority.Bin.Ranking}}{Listing Priority Number (LPN) or priority bin}
\item{\code{Timeframe}}{Fiscal Year for status review}
\item{\code{Current.Candidate}}{Current candidate for listing?}
\item{\code{Range}}{List of states and territories in which species is found}
\item{\code{LPN}}{Listing Priority Number; 1 - 12}
\item{\code{Priority}}{Status review priority bin; 1-5}
}}
\source{
\url{https://www.fws.gov/endangered/esa-library}
}
\usage{
data_clean
}
\description{
The US Fish and Wildlife Service (FWS) provides a seven year plan 
oulining prioritization of pending ESA status reviews, the process the
agency uses to determine whether a species requires federal protection.  
The workplan documents all species with pending status reviews, and
provides information about each species' review timeline and prioritization.
}
\keyword{datasets}

