% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{identify}
\alias{identify}
\title{Identify species based on priority and timeline criteria.}
\usage{
identify(dat, x, y, crit)
}
\arguments{
\item{dat}{A data frame}

\item{x}{Numeric value of priority level}

\item{y}{Numeric indicating fiscal year}

\item{crit}{String indicating "LPN" or "Priority"}
}
\value{
data frame subset of \code{dat}
}
\description{
Identify species based on priority and timeline criteria.
}
\examples{
identify(data, 8, 2020, "LPN")
}

